+++ kairo-slack/src/main/kotlin/kairo/slack/SlackClient.kt
@@ -2,4 +2,7 @@ package kairo.slack
 
 import com.slack.api.methods.AsyncMethodsClient
 
-public typealias SlackClient = AsyncMethodsClient
+public class SlackClient(
+  asyncMethodsClient: AsyncMethodsClient,
+  public val channels: Map<String, String>,
+) : AsyncMethodsClient by asyncMethodsClient